IPluginDrawer接口DrawRectOutLine增加绘制圆角矩形的功能，增加DrawLine函数

+++ TrafficMonitor/DrawCommon.cpp
@@ -1,6 +1,6 @@
 ﻿#include "stdafx.h"
 #include "DrawCommon.h"
-#include "TrafficMonitor.h"
+#include "DrawCommonHelper.h"
 
 CDrawCommon::CDrawCommon()
 {
@@ -16,6 +16,7 @@ void CDrawCommon::Create(CDC* pDC, CWnd* pMainWnd)
     m_pMainWnd = pMainWnd;
     if (pMainWnd != nullptr)
         m_pfont = m_pMainWnd->GetFont();
+    m_gdi_plus_drawer.Create(pDC);
 }
 
 void CDrawCommon::SetFont(CFont* pfont)
@@ -27,6 +28,7 @@ void CDrawCommon::SetFont(CFont* pfont)
 void CDrawCommon::SetDC(CDC* pDC)
 {
     m_pDC = pDC;
+    m_gdi_plus_drawer.Create(pDC);
 }
 
 void CDrawCommon::DrawWindowText(CRect rect, LPCTSTR lpszString, COLORREF color, Alignment align, bool draw_back_ground, bool multi_line, BYTE alpha)
@@ -142,18 +144,26 @@ void CDrawCommon::FillRectWithBackColor(CRect rect)
     m_pDC->FillSolidRect(rect, m_back_color);
 }
 
-void CDrawCommon::DrawRectOutLine(CRect rect, COLORREF color, int width, bool dot_line, BYTE alpha)
+void CDrawCommon::DrawRectOutLine(CRect rect, COLORREF color, int width, bool dot_line, BYTE alpha, int radius)
 {
-    CPen aPen, *pOldPen;
-    aPen.CreatePen((dot_line ? PS_DOT : PS_SOLID), width, color);
-    pOldPen = m_pDC->SelectObject(&aPen);
-    CBrush* pOldBrush{dynamic_cast<CBrush*>(m_pDC->SelectStockObject(NULL_BRUSH))};
+    if (radius > 0)
+    {
+        //半径大于0时绘制圆角矩形，使用GDI+绘制
+        m_gdi_plus_drawer.DrawRectOutLine(rect, color, width, dot_line, alpha, radius);
+    }
+    else
+    {
+        CPen aPen, * pOldPen;
+        aPen.CreatePen((dot_line ? PS_DOT : PS_SOLID), width, color);
+        pOldPen = m_pDC->SelectObject(&aPen);
+        CBrush* pOldBrush{ dynamic_cast<CBrush*>(m_pDC->SelectStockObject(NULL_BRUSH)) };
 
-    rect.DeflateRect(width / 2, width / 2);
-    m_pDC->Rectangle(rect);
-    m_pDC->SelectObject(pOldPen);
-    m_pDC->SelectObject(pOldBrush); // Restore the old brush
-    aPen.DeleteObject();
+        rect.DeflateRect(width / 2, width / 2);
+        m_pDC->Rectangle(rect);
+        m_pDC->SelectObject(pOldPen);
+        m_pDC->SelectObject(pOldBrush); // Restore the old brush
+        aPen.DeleteObject();
+    }
 }
 
 void CDrawCommon::GetRegionFromImage(CRgn& rgn, CBitmap& cBitmap, int threshold)
@@ -201,15 +211,15 @@ int CDrawCommon::GetColorBritness(COLORREF color)
     return (GetRValue(color) + GetGValue(color) + GetBValue(color)) / 3;
 }
 
-void CDrawCommon::DrawLine(CPoint start_point, int height, COLORREF color, BYTE alpha)
+void CDrawCommon::DrawLine(CPoint start_point, CPoint end_point, COLORREF color, BYTE alpha)
 {
     CPen aPen, *pOldPen;
     aPen.CreatePen(PS_SOLID, 1, color);
     pOldPen = m_pDC->SelectObject(&aPen);
     CBrush* pOldBrush{dynamic_cast<CBrush*>(m_pDC->SelectStockObject(NULL_BRUSH))};
 
-    m_pDC->MoveTo(start_point); //移动到起始点，默认是从下向上画
-    m_pDC->LineTo(CPoint(start_point.x, start_point.y - height));
+    m_pDC->MoveTo(start_point);
+    m_pDC->LineTo(end_point);
     m_pDC->SelectObject(pOldPen);
     m_pDC->SelectObject(pOldBrush); // Restore the old brush
     aPen.DeleteObject();
@@ -225,168 +235,4 @@ void CDrawCommon::GetTextExtent(const wchar_t* lpszString, int& w, int& h)
     CSize size = m_pDC->GetTextExtent(lpszString);
     w = size.cx;
     h = size.cy;
-}
-
-UINT DrawCommonHelper::ProccessTextFormat(CRect rect, CSize text_length, IDrawCommon::Alignment align, bool multi_line) noexcept
-{
-    UINT result; // CDC::DrawText()函数的文本格式
-    if (multi_line)
-        result = DT_EDITCONTROL | DT_WORDBREAK | DT_NOPREFIX;
-    else
-        result = DT_VCENTER | DT_SINGLELINE | DT_NOPREFIX;
-
-    if (text_length.cx > rect.Width()) //如果文本宽度超过了矩形区域的宽度，设置了居中时左对齐
-    {
-        if (align == IDrawCommon::Alignment::RIGHT)
-            result |= DT_RIGHT;
-    }
-    else
-    {
-        switch (align)
-        {
-        case IDrawCommon::Alignment::RIGHT:
-            result |= DT_RIGHT;
-            break;
-        case IDrawCommon::Alignment::CENTER:
-            result |= DT_CENTER;
-            break;
-        }
-    }
-    return result;
-}
-
-void DrawCommonHelper::ImageDrawAreaConvert(CSize image_size, CPoint& start_point, CSize& size, IDrawCommon::StretchMode stretch_mode)
-{
-    if (size.cx == 0 || size.cy == 0)       //如果指定的size为0，则使用位图的实际大小绘制
-    {
-        size = CSize(image_size.cx, image_size.cy);
-    }
-    else
-    {
-        if (stretch_mode == IDrawCommon::StretchMode::FILL)
-        {
-            float w_h_ratio, w_h_ratio_draw;        //图像的宽高比、绘制大小的宽高比
-            w_h_ratio = static_cast<float>(image_size.cx) / image_size.cy;
-            w_h_ratio_draw = static_cast<float>(size.cx) / size.cy;
-            if (w_h_ratio > w_h_ratio_draw)     //如果图像的宽高比大于绘制区域的宽高比，则需要裁剪两边的图像
-            {
-                int image_width;        //按比例缩放后的宽度
-                image_width = image_size.cx * size.cy / image_size.cy;
-                start_point.x -= ((image_width - size.cx) / 2);
-                size.cx = image_width;
-            }
-            else
-            {
-                int image_height;       //按比例缩放后的高度
-                image_height = image_size.cy * size.cx / image_size.cx;
-                start_point.y -= ((image_height - size.cy) / 2);
-                size.cy = image_height;
-            }
-        }
-        else if (stretch_mode == IDrawCommon::StretchMode::FIT)
-        {
-            CSize draw_size = image_size;
-            float w_h_ratio, w_h_ratio_draw;        //图像的宽高比、绘制大小的宽高比
-            w_h_ratio = static_cast<float>(image_size.cx) / image_size.cy;
-            w_h_ratio_draw = static_cast<float>(size.cx) / size.cy;
-            if (w_h_ratio > w_h_ratio_draw)     //如果图像的宽高比大于绘制区域的宽高比
-            {
-                draw_size.cy = draw_size.cy * size.cx / draw_size.cx;
-                draw_size.cx = size.cx;
-                start_point.y += ((size.cy - draw_size.cy) / 2);
-            }
-            else
-            {
-                draw_size.cx = draw_size.cx * size.cy / draw_size.cy;
-                draw_size.cy = size.cy;
-                start_point.x += ((size.cx - draw_size.cx) / 2);
-            }
-            size = draw_size;
-        }
-    }
-}
-
-void DrawCommonHelper::GetBitmapAlphaPixel(HBITMAP hBitmap, std::set<Point>& points)
-{
-    points.clear();
-    BITMAP bm;
-    GetObject(hBitmap, sizeof(BITMAP), &bm);
-
-    int width = bm.bmWidth;
-    int height = bm.bmHeight;
-
-    // 获取位图的像素数据
-    BITMAPINFO bmpInfo = { 0 };
-    bmpInfo.bmiHeader.biSize = sizeof(BITMAPINFOHEADER);
-    bmpInfo.bmiHeader.biWidth = width;
-    bmpInfo.bmiHeader.biHeight = -height; // top-down DIB
-    bmpInfo.bmiHeader.biPlanes = 1;
-    bmpInfo.bmiHeader.biBitCount = 32;
-    bmpInfo.bmiHeader.biCompression = BI_RGB;
-
-    HDC hdc = CreateCompatibleDC(NULL);
-    SelectObject(hdc, hBitmap);
-
-    // 分配内存存储位图像素
-    RGBQUAD* pPixels = new RGBQUAD[width * height];
-    GetDIBits(hdc, hBitmap, 0, height, pPixels, &bmpInfo, DIB_RGB_COLORS);
-
-    // 遍历所有像素点
-    for (int y = 0; y < height; ++y)
-    {
-        for (int x = 0; x < width; ++x)
-        {
-            int index = y * width + x;
-            //添加alpha值为0的像素点
-            if (pPixels[index].rgbReserved == 0)
-                points.insert(Point(x, y));
-        }
-    }
-
-    delete[] pPixels;
-    DeleteDC(hdc);
-
-}
-
-void DrawCommonHelper::FixBitmapTextAlpha(HBITMAP hBitmap, BYTE alpha, std::set<Point> alpha_points)
-{
-    BITMAP bm;
-    GetObject(hBitmap, sizeof(BITMAP), &bm);
-
-    int width = bm.bmWidth;
-    int height = bm.bmHeight;
-
-    // 获取位图的像素数据
-    BITMAPINFO bmpInfo = { 0 };
-    bmpInfo.bmiHeader.biSize = sizeof(BITMAPINFOHEADER);
-    bmpInfo.bmiHeader.biWidth = width;
-    bmpInfo.bmiHeader.biHeight = -height; // top-down DIB
-    bmpInfo.bmiHeader.biPlanes = 1;
-    bmpInfo.bmiHeader.biBitCount = 32;
-    bmpInfo.bmiHeader.biCompression = BI_RGB;
-
-    HDC hdc = CreateCompatibleDC(NULL);
-    SelectObject(hdc, hBitmap);
-
-    // 分配内存存储位图像素
-    RGBQUAD* pPixels = new RGBQUAD[width * height];
-    GetDIBits(hdc, hBitmap, 0, height, pPixels, &bmpInfo, DIB_RGB_COLORS);
-
-    // 遍历所有像素
-    for (int y = 0; y < height; ++y)
-    {
-        for (int x = 0; x < width; ++x)
-        {
-            int index = y * width + x;
-            //如果检测到alpha值为0，但是却不在alpha_points里，将其修正为正确的alpha值
-            if (pPixels[index].rgbReserved == 0 && !alpha_points.contains(Point(x, y)))
-                pPixels[index].rgbReserved = alpha; // 设置Alpha通道
-        }
-    }
-
-    // 将修改后的像素数据写回位图
-    SetDIBits(hdc, hBitmap, 0, height, pPixels, &bmpInfo, DIB_RGB_COLORS);
-
-    delete[] pPixels;
-    DeleteDC(hdc);
 }
